BACKPORT: ILP32: arm64: compat ABI: use non-compat openat and open_by_handle_at variants
The only difference is that non-compat version forces O_LARGEFILE,
and it should be the default behaviour for all architectures, as
we are going to drop the support of 32-bit userspace off_t.

The exception is tile32 that continues with compat version of syscalls.

This patch is required to support the new consensus ILP32 implementation.
The patch is ported from https://github.com/norov/linux/commit/7a8fdaf90277758af6d268d637cb5d2c5d7249c6

Change-Id: Ic903f3b05e8cb5c923be6e463c2cdee015fa7ead
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
Acked-by: Arnd Bergmann <arnd@arndb.de>
Acked-by: Chris Metcalf <cmetcalf@ezchip.com> [for tile]
Signed-off-by: Van Duc Uy <Uy.Van@amperecomputing.com>