BACKPORT: usb:xhci: update condition to select bus->sysdev from parent device
In commit 4c39d4b949d36faffbc726525b469886cf311d1c, sysdev points to
devices known to the system firmware or hardware for DMA parameters.
However, the parent of the system firmware/hardware device checking logic does
not work in ACPI boot mode. This patch updates the formulation
to check this case in both DT and ACPI.

This patch is required to properly support USB on Ampere eMag SoC's.

This patch is backported from:
https://www.spinics.net/lists/linux-usb/msg157781.html

Change-Id: Ia657a0ae7f28014c9b2df0218608f65eaa42cb11
Signed-off-by: Tung Nguyen <tunguyen@apm.com>
Signed-off-by: Thang Q. Nguyen <tqnguyen@apm.com>
Signed-off-by: Khuong Dinh <kdinh@apm.com>