BACKPORT: ILP32: arm64: signal: Make parse_user_sigframe() independent of rt_sigframe layout
ILP32 support uses the same struct sigcontext as the native ABI (i.e.,
LP64), but a different layout for the rest of the signal frame (since
siginfo_t and ucontext_t are both ABI-dependent).

Since the purpose of parse_user_sigframe() is really to parse sigcontext
and not the whole signal frame, the function does not need to depend
on the layout of rt_sigframe -- the only purpose of the rt_sigframe
pointer is for use as a base to measure the signal frame size.

So, this patch renames the function to make it clear that only the
sigcontext is really being parsed, and makes the sigframe base pointer
generic.  A macro is defined to provide a suitable duck-typed interface
that can be used with both sigframe definitions.

This patch is required to support the new consensus ILP32 implementation.
The patch is ported from https://github.com/norov/linux/commit/43bff26616fc97d1ea5615753ec63a11536814d0

Change-Id: If94113effd160390e57ab4320321a0abe2f7d175
Suggested-by: Yury Norov <ynorov@caviumnetworks.com>
Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
Signed-off-by: Van Duc Uy <Uy.Van@amperecomputing.com>