BACKPORT: ACPI: SPCR: Use access width to determine mmio usage
The current SPCR code does not check the access width of the MMIO, and
uses a default of 8bit register accesses.  This prevents devices that
only do 16 or 32bit register accesses from working.  By simply checking
this field and setting the MMIO string appropriately, this issue can be
corrected.  To prevent any legacy issues, the code will default to 8bit
accesses if the value is anything but 16 or 32.

This patch adds serial console support only for X-Gene v1 and v2.
This patch is not required for Ampere eMag SoC's.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?h=next-20171017&id=2bece49394872d36bbc5767fd643deac05920c55

Change-Id: Ie04edc7c96ea4d8864a91105d33728ba961153a1
Signed-off-by: Jon Mason <jon.mason@broadcom.com>
Signed-off-by: Loc Ho <lho@apm.com>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Khuong Dinh <kdinh@apm.com>