BACKPORT: ACPI: SPCR: Workaround for APM X-Gene 8250 UART 32-alignment errata
APM X-Gene version 1 and 2 have an 8250 UART with its register
aligned to 32-bit. In addition, the latest released BIOS
encodes the access field as 8-bit access instead 32-bit access.
This causes no console with ACPI boot as the console
will not match X-Gene UART port due to the lack of mmio32
option.

This patch adds serial console support only for X-Gene v1 and v2.
This is not required for Ampere eMag SoC's.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?h=next-20171017&id=79a648328d2a604524a30523ca763fbeca0f70e3

Change-Id: I041f7243a651ae1172aab712deb11a0768e5b272
Signed-off-by: Loc Ho <lho@apm.com>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Khuong Dinh <kdinh@apm.com>