BACKPORT: ACPI: SPCR: extend XGENE 8250 workaround to m400
xgene v1/v2 chips are also used on moonshot cartridges that have
different table headers to the ones on Mustang. Extend the quirk
so it also recognises the Moonshot M400 variant too.

This patch adds serial console support only for X-Gene v1 and v2.
This is not required for Ampere eMag SoC's.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?h=next-20171017&id=dee82bc1e653126408f4108cd994d1e96949d064

Change-Id: I7a28b189d9da22beb327e05f3afa3d450fc9fdca
Signed-off-by: Graeme Gregory <graeme.gregory@linaro.org>
Tested-by: Mark Salter <msalter@redhat.com>
Reviewed-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Khuong Dinh <kdinh@apm.com>