BACKPORT: ILP32: arm64: ilp32: Make the Kconfig option default y
The intention of the ILP32 branches is to enable ILP32 by default. This
default is to be revisited for upstream merging.

This patch is required to support the new consensus ILP32 implementation.
The patch is ported from https://github.com/norov/linux/commit/88c5962bd33e5c7721b54af19090811128418a09

Change-Id: Iadb8e073ca317a84b0f29174f98a4d55380c46ad
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
Signed-off-by: Van Duc Uy <Uy.Van@amperecomputing.com>