BACKPORT: PCI: Add ACS quirk for APM X-Gene devices
The APM X-Gene PCIe root port does not support ACS at this point.  However,
the hardware provides isolation and source validation through the SMMU.
The stream ID generated by the PCIe ports contain both the bus/device/
function number as well as the port ID in its 3 most significant bits.
Turn on ACS but disable all the peer-to-peer features.

This patch is required to properly support PCIe in Ampere eMag SoC's.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?h=next-20171017&id=a0418aa262dfbe4011df85c8d40993bf21d53cd2

Change-Id: Ibc72a021d7b747aa7e610d32d782995bd1b77a62
Signed-off-by: Feng Kan <fkan@apm.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Alex Williamson <alex.williamson@redhat.com>
Acked-by: Tanmay Inamdar <tinamdar@apm.com>
Signed-off-by: Khuong Dinh <kdinh@apm.com>