BACKPORT: drivers: net: xgene: Correct probe sequence handling
The phy is connected at early stage of probe but not properly
disconnected if error occurs.  This patch fixes the issue.

Also changing the return type of xgene_enet_check_phy_handle(),
since this function always returns success.

This patch is required only for X-Gene SoC v1 and v2.
It is not applicable for Ampere SoC's.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?h=next-20170929&id=183db481279437590f75a8a0479d512e5dd597de

Change-Id: Ide91a21b9a2df9cba635df3039c0f0075a47173e
Signed-off-by: Quan Nguyen <qnguyen@apm.com>
Signed-off-by: Iyappan Subramanian <isubramanian@apm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Khuong Dinh <kdinh@apm.com>