NOUPSTREAM: centos: defconfig: Change CMA_SIZE to 0
This patch disables CMA to let the system use the default swiotlb
DMA allocation.
It is observed certain drivers such as 1 LSI and 2 Mellanox cards
failed to initialize if they failed to allocate CMA buffer. But
they would function properly if there is no CMA buffer to start
with. As a temporary solution, disable CMA by set to 0 until this
is better understood.

This issue is fixed from upstream 4.14 kernel. For more recent kernel,
the patch will not be needed.

Change-Id: I7b1e0b3c942840ab6e370022a1ec67c80c198f93
Signed-off-by: Khuong Dinh <kdinh@apm.com>