BACKPORT: iommu: enable bypass transaction caching for ARM SMMU 500
The ARM SMMU identity mapping performance was poor compared with the
DMA mode. It was found that enable caching would restore the performance
back to normal. The S2CRB_TLBEN bit in the ACR register would allow for
caching of the stream to context register bypass transaction information.

This patch is required to properly support PCIe in virtual environment
for Ampere eMag SoC's.

This patch is backported from: https://patchwork.kernel.org/patch/10000707/
and is under open source review progress.

Change-Id: Iea8230870f7b39d1006c2e515707a0e612e69cc9
Signed-off-by: Feng Kan <fkan@apm.com>
Signed-off-by: Khuong Dinh <kdinh@apm.com>