BACKPORT: iommu: use the smallest DMA aperture to set dma_limit
The current DMA IOVA allocation only check against the smallest
aperture between the driver and IOMMU hardware. In some system
where the IOMMU aperture is greater than the bus width set by
the IOMMU, incorrect IOVA could be used and causes hang and crashes.
Add addition field to record the ACPI _DMA aperture and use the
smallest of the three to set the dma_limit.

This patch comes with the commit cd4437 to resolve iommu_dma_unmap()
crash issue and pass traffic for XL710 and ConnectX-5 for the case of
iommu.passthrough=0.

This patch is required to properly support PCIe in virtual environment
for Ampere eMag SoC's.

This patch is under upstream discussion of DMA aperture. The discussion
can be found at https://lkml.org/lkml/2017/12/1/245.

Change-Id: I9fbfda485f30ca59c7ad5e9a9f92387a16ebbab2
Signed-off-by: Feng Kan <fkan@apm.com>
Signed-off-by: Toan Le <toanle@apm.com>