BACKPORT: arm64: cpufeature: __this_cpu_has_cap() shouldn't stop early
this_cpu_has_cap() tests caps->desc not caps->matches, so it stops
walking the list when it finds a 'silent' feature, instead of
walking to the end of the list.

Prior to v4.6's 644c2ae198412 ("arm64: cpufeature: Test 'matches' pointer
to find the end of the list") we always tested desc to find the end of
a capability list. This was changed for dubious things like PAN_NOT_UAO.
v4.7's e3661b128e53e ("arm64: Allow a capability to be checked on
single CPU") added this_cpu_has_cap() using the old desc style test.

Without this patch, CPU can't be hot plugged.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?h=v4.18.12&id=edf298cfce47ab7279d03b5203ae2ef3a58e49db

Change-Id: I140dbe05f44ab6d7ce8a94c4a8576692ba86fc1e
CC: Suzuki K Poulose <suzuki.poulose@arm.com>
Reviewed-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>