BACKPORT: blk-mq: update ->init_request and ->exit_request prototypes
Remove the request_idx parameter, which can't be used safely now that we
support I/O schedulers with blk-mq. Except for a superflous check in
mtip32xx it was unused anyway.

Also pass the tag_set instead of just the driver data - this allows drivers
to avoid some code duplication in a follow on cleanup.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jens Axboe <axboe@fb.com>

The patch "scsi: Avoid that scsi_exit_rq() triggers a use-after-free" is on top of this patch.
This patch is required as part of the resolution to the USB issue when hot remove
of the USB storage device. Otherwise, the kernel will panic with below back trace:

[<ffff0000081f4b98>] kmem_cache_free+0x20/0x168
[<ffff0000084c0e30>] scsi_exit_rq+0x60/0x70
[<ffff000008312198>] free_request_size+0x28/0x40
[<ffff000008199e1c>] mempool_destroy+0x2c/0x78
[<ffff000008313e88>] blk_exit_rl+0x20/0x40
[<ffff000008317a78>] __blk_release_queue+0x50/0xe8
[<ffff0000080d7298>] process_one_work+0x178/0x338
[<ffff0000080d74a4>] worker_thread+0x4c/0x438
[<ffff0000080dd8d4>] kthread+0xfc/0x128
[<ffff000008082bc0>] ret_from_fork+0x10/0x50

This patch is required to support hotplug for USB. Without this patch,
USB hotplug will cause kernel crash.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v4.13&id=d6296d39e90c9075bc2fc15f1e86dac44930d4b5

Change-Id: Ia4cbe65b38562722b3f50c82bb670e2979d6a04a
Signed-off-by: Tung Nguyen <tunguyen@apm.com>