BACKPORT: perf: Add Arm CMN-600 DT binding
Document the requirements for the CMN-600 DT binding. The internal
topology is almost entirely discoverable by walking a tree of ID
registers, but sadly both the starting point for that walk and the
exact format of those registers are configuration-dependent and not
discoverable from some sane fixed location. Oh well.

This patch is part of the eMAG2 CMN-600 PMU patch set.
This patch is backported from:
http://www.linux-arm.org/git?p=linux-rm.git;a=commit;h=d92b2ecc303fba919a57d699ae4fd297ced2e91c

Signed-off-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>