BACKPORT: PM / core: fix deferred probe breaking suspend resume order
When bridge and its endpoint is enumerated the devices are added to the
dpm list. Afterward, the bridge defers probe when IOMMU is not ready.
This causes the bridge to be moved to the end of the dpm list when
deferred probe kicks in. The order of the dpm list for bridge and
endpoint is reversed.

Add reordering code to move the bridge and its children and consumers to
the end of the pm list so the order for suspend and resume is not altered.
The code also move device and its children and consumers to the tail of
device_kset list if it is registered.

This patch is required to correct suspend and resume of PCIe hierarchy link
including root complexes and downstream endpoint devices. On the system
with ARM/SMMUv2, deferred probe happening on root complex will cause its
moving to the end of device list. Individually moving root complex breaks
the dependency to its downstream endpoint; hence, it breaks the suspend and
resume. This patch fixes by moving the whole chain to the end of the list.

This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm.git/commit/?h=linux-next&id=494fd7b7ad10c33d3a7ff7d10b71b3ecad10474a

Change-Id: Ib700afa25dbbb5f2fd9a1584fd590f9a44fdc3f4
Signed-off-by: Feng Kan <fkan@apm.com>
Signed-off-by: Toan Le <toanle@apm.com>
Reviewed-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>