BACKPORT: PCI: Add ACS quirk for Ampere root ports
The Ampere Computing PCIe root port does not support ACS at this point.
However, the hardware provides isolation and source validation through the
SMMU. The stream ID generated by the PCIe ports contain both the
bus/device/function number as well as the port ID in its 3 most significant
bits. Turn on ACS but disable all the peer-to-peer features.

APM is being rebranded to Ampere. The Vendor and Device IDs change,
but the functionality stays the same.

This patch is needed to maintain PCIe virtualization support with Ampere
Vendor ID. This patch does not break backward compatibility with former
AppiedMicro Vendor ID.

The upstream patch is in 4.17-rc1.
This patch is backported from:
https://git.kernel.org/pub/scm/linux/kernel/git/helgaas/pci.git/commit/?id=4ef76ad0462cf25ce948541c8724eaa8a8365e1d

Change-Id: I0a88b3b634f73baa058c15255908f01162f90cac
Signed-off-by: Feng Kan <fkan@apm.com>
Signed-off-by: Bjorn Helgaas <helgaas@kernel.org>
[toanle to back-port to Ampere CentOS tree]
Signed-off-by: Toan Le <toanle@apm.com>