BACKPORT: VARIANT 2: arm64: Add initial retpoline support.
Enable the use of -mindirect-branch=thunk-extern in newer GCC, and
provide the corresponding thunks. Provide assembler macros for invoking the
thunks in the same way that GCC does, from native and inline assembler.
The integration is inspired by the x86 retpoline support.

This patch is part of the eMAG Spectre variant 2 patch set.
It can be found at:
https://git.theobroma-systems.com/ampere-computing/linux.git/commit/?h=arm64-for-upstream-24534b35&id=69610a2b7c7512f4e951d05b4fdf86b2e7d46770

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>