BACKPORT: VARIANT 2: arm64: Use alternative framework for retpoline.
This reduces the impact of retpoline on non-affected processors.
By default the retpoline will consist of 6 nop instructions.
In case an affected processor is detected (by a MIDR match)
the nops will be replaced by the retpoline.

The processors, which are using this are Applied Micro's X-Gene1, X-Gene2,
X-Gene3, and Ampere Computing's eMAG1.

This patch is part of the eMAG Spectre variant 2 patch set.
It can be found at:
https://git.theobroma-systems.com/ampere-computing/linux.git/tree/arch/arm64/kernel/cpu_errata.c?h=arm64-for-upstream-24534b35&id=477b9791f15ad66ac25e308530bcc6a2556da967

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>