BACKPORT: perf: xgene: Add CPU hotplug support
This patch adds CPU hotplug support where the PMU migrates the context to
another online CPU when its CPU is offline.

It fixes the below issue where the user does offline the CPU which is assigned
to this PMU.

Assuming, CPU0 is assigned for this PMU. When the user does offline CPU0
	[root@(none) ~]# echo 0 > /sys/devices/system/cpu/cpu0/online
This PMU does not work anymore and shows the below error.
	[root@(none) ~]# perf stat -a -e l3c0/cycle-count/,l3c0/write/ sleep 1
	Error:
	The sys_perf_event_open() syscall returned with 19 (No such device) for event (l3c0/cycle-count/).
	/bin/dmesg may provide additional information.
	No CONFIG_PERF_EVENTS=y kernel support configured?

With this patch, when CPU0 is offline, PMU migrates to another online CPU and
works on that CPU.

This patch is referred to the posted patch version 1
https://patchwork.kernel.org/patch/10566739/

Given that this patch is still in-flight, it will be updated in the future
with an accepted version.

Change-Id: I4f54a9210e0f37cf58664c7ffc5ab0bafb854487
Signed-off-by: Hoan Tran <hoan.tran@amperecomputing.com>